q4 code

+++ EC330/HW4/q4.cpp
@@ -106,7 +106,7 @@ vector<string> brute_force_5(vector<string> hashes){
 		cracks.push_back("NONE_FOUND");
 
 	int distance = END_cHAR - START_CHAR + 1;
-	double total_iters = distance*distance*distance*distance*distance;
+	double total_iters = (float)distance*(float)distance*(float)distance*(float)distance*(float)distance;
 	double current_iteration = 0;
 	double percentage = total_iters / 10000, percentage_val = percentage;
 	printf("%f\n", total_iters);
